auth/pyproject.toml
[project]
name = "jarvis-auth-service"
version = "0.1.0"
description = "Authentication service for Jarvis"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.115.13",
    "sqlalchemy>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    # https://github.com/pyca/bcrypt/issues/684
    "bcrypt==4.0.1",
    "python-multipart>=0.0.6",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.16.4",
]